Move docs search into header
